+++ src/App.jsx
@@ -4,6 +4,7 @@ import "./index.css";
 import useApi from "./Hooks/useApi";
 import fetchData from "./Utils/Functions/fetchInformation";
 import Search from "./Utils/Components/Search";
+import { PaginationFooter } from "./Utils/Components/PaginationFooter";
 
 function App() {
   const [page, setPage] = useState(1);
@@ -12,24 +13,6 @@ function App() {
   const [routes, setRoutes] = useState([]);
   const [search, setSearch] = useState("");
   const api = useApi();
-
-  /***common for all pages for pagination */
-
-  const createPaginationLinks = (lastPage, currentPage, setCurrentPage) => {
-    const links = Array.from({ length: lastPage }).map((_, i) => (
-      <li
-        key={i + 1}
-        className={`page-item ${i + 1 === currentPage ? "active" : ""}`}
-      >
-        <button className="page-link" onClick={() => setCurrentPage(i + 1)}>
-          {i + 1}
-        </button>
-      </li>
-    ));
-    return links;
-  };
-
-  /***common for all pages for paginate */
   const fetchRouteInformation = async () => {
     await fetchData( api.fetchRoutes, page,setLastPage,setRoutes, search, setPaginationInformation);
   };
@@ -73,7 +56,7 @@ function App() {
                       </thead>
                       <tbody>
                         {routes.map((route, index) => (
-                          <tr key={route.id}>
+                          <tr key={index}>
                             <td>{route.id}</td>
                             <td>{route.origin}</td>
                             <td>
@@ -86,35 +69,7 @@ function App() {
                       </tbody>
                     </table>
                   </div>
-                  <div className="card-footer d-flex align-items-center">
-                    <p className="m-0 text-muted">
-                      Showing <span>{paginationInformation.from}</span> to <span>{paginationInformation.to}</span> of{" "}
-                      <span>{paginationInformation.total}</span> entries
-                    </p>
-                    <ul className="pagination m-0 ms-auto">
-                      {createPaginationLinks(lastPage, page, setPage)}
-                      <li className="page-item">
-                        <a className="page-link" href="#">
-                          next
-                          <svg
-                            xmlns="http://www.w3.org/2000/svg"
-                            className="icon"
-                            width={24}
-                            height={24}
-                            viewBox="0 0 24 24"
-                            strokeWidth={2}
-                            stroke="currentColor"
-                            fill="none"
-                            strokeLinecap="round"
-                            strokeLinejoin="round"
-                          >
-                            <path stroke="none" d="M0 0h24v24H0z" fill="none" />
-                            <path d="M9 6l6 6l-6 6" />
-                          </svg>
-                        </a>
-                      </li>
-                    </ul>
-                  </div>
+                  <PaginationFooter paginationInformation={paginationInformation} lastPage={lastPage} page={page} setPage={setPage} />
                 </div>
               </div>
             </div>
